Add SDF-focused training scripts and SuperSplat conversion tools
- Add train_sdf_simple.sh: SDF几何优化专用脚本，禁用PBR复杂度
- Add train_complex_scene.sh: 复杂场景稳定训练参数配置
- Add convert_to_supersplat.py: 2D高斯转3D高斯，支持SuperSplat实时浏览
- Add run_simple.sh: 简化训练流程脚本
- Fix dataset_readers.py: mask数组类型转换修复
- Add CLAUDE.md: 项目配置和使用说明文档

保留SDF几何优化核心功能，移除PBR材质复杂度，提升复杂场景收敛稳定性

+++ scene/dataset_readers.py
@@ -120,7 +120,7 @@ def readColmapCameras(cam_extrinsics, cam_intrinsics, images_folder):
             else:
                 mask = np.ones_like(np.array(image)[:, :, 0]) * 255.
             image = Image.fromarray(image)
-            mask = Image.fromarray(mask)
+            mask = Image.fromarray(mask.astype(np.uint8))
             mask.save('test.png')
 
         cam_info = CameraInfo(uid=uid, R=R, T=T, FovY=FovY, FovX=FovX, image=image,
